+++ packages/integration-sdk-entity-validator/src/j1Formats.ts
@@ -9,9 +9,12 @@ const ipv4CidrRegex =
 const ipv6CidrRegex =
   /^([0-9a-fA-F]{1,4}:){7,7}[0-9a-fA-F]{1,4}\/(12[0-8]|1[01][0-9]|[1-9]?[0-9])$|^([0-9a-fA-F]{1,4}:){1,7}:\/(12[0-8]|1[01][0-9]|[1-9]?[0-9])$|^::\/(12[0-8]|1[01][0-9]|[1-9]?[0-9])$|^([0-9a-fA-F]{1,4}:){1,7}[0-9a-fA-F]{1,4}$/;
 
+const asnRegex = /^AS\d{2,6}$/;
+
 const j1JsonSchemaFormats = {
   ip: (x) => ipv4.test(x) || ipv6.test(x),
   ipCidr: (x) => ipv4CidrRegex.test(x) || ipv6CidrRegex.test(x),
+  asn: (x: string) => asnRegex.test(x),
 } satisfies Record<string, AjvFormat>;
 
 export const addJ1Formats = (ajvInstance: Ajv) => {